Created a Cell Shader
I've create a Windwaker style Cell Shader to use for our game, to make all of the models look "Cartoony". It can only interact with one directional light but it also allows for the shadow size, color and blend to be tweaked. It also has a Rim light with the same features and it can also be turned off.