Environment shader
The environment shader should be able to recieve shadows but i have no idea how to get them to be a flat as the ones it casts on itself using the current shader. i'm not entirely sure how this approach will work and i won't focus to hard on it. Also added Pro Builder.